requirements.txt
Flask==2.0.3
numpy
pandas
-f https://download.pytorch.org/whl/torch_stable.html 
torch==1.13.1+cpu
torchvision==0.14.1+cpu
pillow
